requirements.txt


Django>=1.8,<2.0
djangorestframework==3.5.3
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3



coverage==4.2
pytest==3.0.3
pytest-django==3.0.0
pytest-cov==2.5.1
django_dynamic_fixture==1.9.5
Faker==0.7.3
pdbpp==0.9.1
fpdf==1.7.2